requirements.txt
pandas
numpy
streamlit
scikit-learn
plotly
matplotlib
seaborn
psycopg2
python-dotenv
seaborn